refactor(gate): DELETE the slot limiter rather than fix it, and rewrite the merge-gate policy to say what is true
Re-scope, on the operator's direction, after applying the algorithm to the
process this branch was optimizing rather than to the branch.

🔴 THE REQUIREMENT THE LIMITER SERVED IS BEING DELETED. "Run both tiers locally
before merging" is removed from CLAUDE.md. Branch protection stays off — a
knowing trade of safety for speed. CI stays advisory. The inner loop becomes a
change-scoped run (a separate PR). A limiter that bounds concurrent full-suite
gate runs is a step-3/4 optimization of a step steps 1 and 2 say should not
exist, so it is deleted, not fixed:

  - `acquire_slot` and its whole SLOT LIMITER comment block
  - DEVRC_GATE_SLOTS / _SLOT_WAIT / _SLOT_DIR / _SLOT_POOL and their --help entries
  - GATE_SLOT_HELD / _FD / _STATE and the `slot:` line in the GATE report
  - the fd-close subshell in `run_tier` — it existed ONLY to stop an orphan
    inheriting the slot lock, so with no lock it has nothing to protect and
    `run_tier` returns to the plain form
  - scripts/tests/test_gate_slots_and_reexec.py -> scripts/tests/test_gate_reexec.py,
    keeping only the re-exec, --help-derivation and ambient-scrub tests

🔴 THE 3.4x CONCURRENCY TABLE IS DELETED, NOT RESTATED. It existed only to
justify the limiter. (For the record, since the branch quoted it as measured
fact: bucketing 237 runs by how many others overlapped them is length-biased —
overlap count is partly a function of the run's own duration — and a null Monte
Carlo with ZERO interaction reproduced the monotone shape, the 14.5 min baseline
and ~2.06x of the 3.4x. Deleting beats correcting here because nothing left in
the tree needs the number.)

What survives, each on its own merit:

1. cgroup-quota worker sizing in run-tests.sh — a correctness fix. `nproc`
   reports the NODE's cores, not the container's.
2. the `nix develop` re-exec in gate.sh — it removes a whole retype cycle from
   26% of gate invocations (101 of 382; the naive "100 of 100" was a population
   selected by the failure's own cause, since the default LOG_DIR is a
   `mktemp -d` and only runs OUTSIDE the dev shell land in bare /tmp).

CLAUDE.MD: the gate policy now says what is true rather than what was expected.
The local full-suite mandate is deleted WITH ITS REASON — it produced 27–50
concurrent full-suite runs on one 24-core box, the same ~22,000 tests twice per
change, gating nothing, and contended badly enough that the dev-host tier
repeatedly hit its own 3600s cap and produced no verdict at all. CI is named as
ADVISORY and as the only automated signal there is. And the bullet says plainly
that this arrangement is LESS SAFE and deliberately so, rather than implying the
rearrangement has made it safe. The `<!-- merge-gate: other -->` marker is
unchanged and still sits on a line that says something to a human.

⚠ ONE MUTANT FOUND AND FIXED IN THIS ROUND, and it is the fixture-on-its-own-
boundary trap: `test_a_quota_wider_than_nproc_does_not_widen_the_budget`
SURVIVED a mutant that let the cgroup quota win outright, because on a 24-core
box `min(nproc, 8)` is 8 and the cap swallowed the wrong answer. The test now
narrows the CHILD's affinity mask to 2 CPUs and uses a quota of 5, so the wrong
answer sits strictly between the right one and the cap — plus a positive control
that fails if the narrowing ever stops taking effect.

⚠ ONE MUTANT STILL SURVIVES AND IS LABELLED, unchanged from the previous commit:
deleting the `q=""; p=""` reset in the cgroup walk kills no test, because bash's
`read` assigns even at EOF and the walk returns on the first numeric quota. It
stays as an invariant guard and both the code comment and the test docstring say
so.

Verified: mutation sweep re-run over the reduced scope — every guard killed by
its own test except the labelled one above. test_run_tests_jobs.py +
test_gate_reexec.py green at nproc = 4, 8 and 24 (taskset). Merged onto
origin/main a2b74e4b. The gating tier is the arbiter for the blocker: on the
previous head a8d2aa01 `tekton/devrc-pytests` reported
`collected=21365 passed=21357 failed=1`, and the five failures in
test_run_tests_jobs.py are GONE — the one remaining is
`test_mjs_parses[attachments.mjs]`, in a file this branch does not touch and
which `node --check` parses clean locally.
